+++ changelog.md
@@ -2,6 +2,36 @@
 
 This changelog only includes the most important changes in recent updates. For a full log of all changes, please refer to git.
 
+### Version 5.1.0
+* Added operator support for integrators SABA, EOS, leapfrog, and BS
+* Fixed bug in operator timestep for adaptive integrators (IAS15, BS)
+* Output from IAS15 with operators will differ slightly from prev versions
+
+### Version 5.0.1 
+* fixed bug saving binaries with tides_spin
+* removed %matplotlib inline from notebooks
+
+### Version 5.0.0
+* Updated to match REBOUND 5.0.0.
+* New string datatype, replaces hashes. 
+* Added free_memory function pointers to force and operator structs to clean up memory they allocate
+* Updated build process and Github CI runners
+
+### Version 4.6.2
+* Updated to require REBOUND < 5.0
+
+### Version 4.6.1
+* Updated to match REBOUND 4.6.0 (merging of internal part1 and part2 functions in integrators)
+
+### Version 4.6.0
+* Added tides\_dynamical to model dynamical tides following Vick \& Lai 2019 (see Liveoak et al. 2025)
+
+### Version 4.5.0
+* Fixed factor of two error in conservative piece of the tidal potential in Eggleton+98 in tides_spin
+
+### Version 4.4.2
+* Fixed bug when resetting ODEs in tides_spin
+
 ### Version 4.4.1
 * Added support for uv installation